# src/main/java/edu/hei/school/restaurant/dtos/responses/IngredientBasicProperty.java
package edu.hei.school.restaurant.dtos.responses;

import lombok.AllArgsConstructor;
import lombok.Getter;

@AllArgsConstructor
@Getter
public class IngredientBasicProperty {
    private Long id;
    private String name;
}